Render each statement on their own row
