+++ app/javascript/components/action_wrapper.html
@@ -1,0 +1,1 @@
+<span v-bind:is="currentView" :statement="statement" :page="page" :country="country"></span>
