+++ app/javascript/components/action_wrapper.js
@@ -1,0 +1,27 @@
+import template from './action_wrapper.html'
+
+import verifiableComponent from './verifiable'
+import unverifiableComponent from './unverifiable'
+import reconcilableComponent from './reconcilable'
+import actionableComponent from './actionable'
+import manuallyActionableComponent from './manually_actionable'
+import doneComponent from './done'
+import revertedComponent from './reverted'
+
+export default template({
+  data () { return { } },
+  props: ['statement', 'page', 'country'],
+  computed: {
+    currentView: function () {
+      switch (this.statement.type) {
+        case 'verifiable': return verifiableComponent
+        case 'unverifiable': return unverifiableComponent
+        case 'reconcilable': return reconcilableComponent
+        case 'actionable': return actionableComponent
+        case 'manually_actionable': return manuallyActionableComponent
+        case 'done': return doneComponent
+        case 'reverted': return revertedComponent
+      }
+    }
+  }
+})
